+++ examples/basic/internlm/basic_language_model_InternLM_tools.py
@@ -20,13 +20,40 @@ generation_config = {
     'top_k': None,
 }
 
+
 functions = [
     {
-        "name": "interpreter", 
-        "description": "你现在已经能够在一个有状态的 Jupyter 笔记本环境中运行 Python 代码。当你向 python 发送含有 Python 代码的消息时，它将在该环境中执行。这个工具适用于多种场景，如数据分析或处理（包括数据操作、统计分析、图表绘制），复杂的计算问题（解决数学和物理难题），编程示例（理解编程概念或特性），文本处理和分析（比如文本解析和自然语言处理），机器学习和数据科学（用于展示模型训练和数据可视化），以及文件操作和数据导入（处理CSV、JSON等格式的文件）。",
+        'name': 'IPythonInterpreter',
+        'description': "When you send a message containing Python code to python, it will be executed in a stateful Jupyter notebook environment. python will respond with the output of the execution or time out after 60.0 seconds. The drive at '/mnt/data' can be used to save and persist user files. Internet access for this session is disabled. Do not make external web requests or API calls as they will fail.",
+        'parameters': [{
+            'name': 'command',
+            'type': 'STRING',
+            'description': 'Python code'
+        }, {
+            'name': 'timeout',
+            'type': 'NUMBER',
+            'description': 'Upper bound of waiting time for Python script execution.'
+        }],
+        'required': ['command'],
+        'parameter_description': 'If you call this tool, you must pass arguments in the JSON format {key: value}, where the key is the parameter name.'
+    },
+    {
+        'name': 'ArxivSearch.get_arxiv_article_information',
+        'description': 'Run Arxiv search and get the article meta information.',
+        'parameters': [{
+            'name': 'query',
+            'type': 'STRING',
+            'description': 'the content of search query'
+        }],
+        'required': ['query'],
+        'return_data': [{
+            'name': 'content',
+            'description': 'a list of 3 arxiv search papers',
+            'type': 'STRING'
+        }],
+        'parameter_description': 'If you call this tool, you must pass arguments in the JSON format {key: value}, where the key is the parameter name.'
     }
 ]
-
 
 cli_demo = Chat(model_dir, 
                 system='当开启工具以及代码时，根据需求选择合适的工具进行调用',